+++ backend/src/http/login_credentials/mod.rs
@@ -13,6 +13,8 @@ use server::State;
 use token::Token;
 use webapp::protocol::{request, response};
 
+mod tests;
+
 pub fn login_credentials<T>(http_request: &HttpRequest<State<T>>) -> FutureResponse
 where
     T: Actor + Handler<CreateSession>,
@@ -48,59 +50,4 @@ where
                 })
         })
         .responder()
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use actix_web::test::TestServer;
-    use database::DatabaseError;
-    use http::tests::{execute_request, state, DatabaseExecutorMock};
-    use serde_cbor::to_vec;
-    use webapp::protocol::{model::Session, request};
-
-    impl Handler<CreateSession> for DatabaseExecutorMock {
-        type Result = Result<Session, DatabaseError>;
-        fn handle(&mut self, _: CreateSession, _: &mut Self::Context) -> Self::Result {
-            Ok(Session {
-                token: Token::create("username").unwrap(),
-            })
-        }
-    }
-
-    fn create_testserver() -> TestServer {
-        TestServer::build_with_state(state).start(|app| app.handler(login_credentials))
-    }
-
-    #[test]
-    fn succeed_to_login_with_credentials() {
-        // Given
-        let mut server = create_testserver();
-        let body = to_vec(&request::LoginCredentials {
-            username: "username".to_owned(),
-            password: "username".to_owned(),
-        }).unwrap();
-
-        // When
-        let response = execute_request(&mut server, body);
-
-        // Then
-        assert!(response.status().is_success());
-    }
-
-    #[test]
-    fn fail_to_login_with_wrong_credentials() {
-        // Given
-        let mut server = create_testserver();
-        let body = to_vec(&request::LoginCredentials {
-            username: "username".to_owned(),
-            password: "password".to_owned(),
-        }).unwrap();
-
-        // When
-        let response = execute_request(&mut server, body);
-
-        // Then
-        assert_eq!(response.status().is_success(), false);
-    }
 }
