+++ backend/src/http/logout/mod.rs
@@ -1,4 +1,4 @@
-//! The session based login request
+//! The logout request
 
 use actix::{dev::ToEnvelope, prelude::*};
 use actix_web::{AsyncResponder, HttpRequest, HttpResponse};
@@ -8,6 +8,8 @@ use futures::Future;
 use http::FutureResponse;
 use server::State;
 use webapp::protocol::{model::Session, request, response};
+
+mod tests;
 
 pub fn logout<T: Actor>(http_request: &HttpRequest<State<T>>) -> FutureResponse
 where
@@ -31,40 +33,4 @@ where
                 })
         })
         .responder()
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use actix_web::test::TestServer;
-    use database::DatabaseError;
-    use http::tests::{execute_request, state, DatabaseExecutorMock};
-    use serde_cbor::to_vec;
-    use webapp::protocol::{model::Session, request};
-
-    impl Handler<DeleteSession> for DatabaseExecutorMock {
-        type Result = Result<(), DatabaseError>;
-        fn handle(&mut self, _: DeleteSession, _: &mut Self::Context) -> Self::Result {
-            Ok(())
-        }
-    }
-
-    fn create_testserver() -> TestServer {
-        TestServer::build_with_state(state).start(|app| app.handler(logout))
-    }
-
-    #[test]
-    fn succeed_to_logout() {
-        // Given
-        let mut server = create_testserver();
-        let body = to_vec(&request::Logout(Session {
-            token: "any-token".to_owned(),
-        })).unwrap();
-
-        // When
-        let response = execute_request(&mut server, body);
-
-        // Then
-        assert!(response.status().is_success());
-    }
 }
