+++ backend/src/token/mod.rs
@@ -4,6 +4,8 @@ use jsonwebtoken::{decode, encode, Header, Validation};
 use time::get_time;
 use uuid::Uuid;
 
+mod tests;
+
 const SECRET: &[u8] = b"my_secret";
 
 #[derive(Debug, Fail)]
@@ -51,26 +53,5 @@ impl Token {
     pub fn verify(token: &str) -> Result<String, TokenError> {
         let data = decode::<Token>(token, SECRET, &Validation::default()).map_err(|_| TokenError::Verify)?;
         Self::create(&data.claims.sub)
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn succeed_to_create_a_token() {
-        assert!(Token::create("").is_ok());
-    }
-
-    #[test]
-    fn succeed_to_verify_a_token() {
-        let sut = Token::create("").unwrap();
-        assert!(Token::verify(&sut).is_ok());
-    }
-
-    #[test]
-    fn fail_to_verify_a_wrong_token() {
-        assert!(Token::verify("wrong").is_err());
     }
 }
